requirements.txt
yfinance>=0.2.40
pandas>=2.0.0
numpy>=1.26.0
requests>=2.31.0
ccxt>=4.3.0
fredapi>=0.5.1
fastapi>=0.110.0
uvicorn>=0.29.0
pyyaml>=6.0
sqlalchemy>=2.0.0
pyzmq>=25.0.0
structlog>=24.0.0
httpx>=0.27.0
python-dotenv>=1.0.0
pytest>=8.0.0
pytest-asyncio>=0.23.0
python-telegram-bot>=21.0
APScheduler>=3.10.0
hmmlearn>=0.3.2
# OSINT intelligence layer
edgartools>=3.0.0
gdeltdoc>=1.5.0
instructor>=1.7.0
websockets>=12.0
